Move certificate profiles logic in separate service part
